pyproject.toml
[project]
name = "learn-computer-vision-in-30-days"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "mediapipe==0.10.9",
    "numpy<2",
    "opencv-python<4.10",
    "pillow>=12.1.0",
    "protobuf<4",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "skl2onnx>=1.20.0",
    "streamlit>=1.53.1",
    "ultralytics>=8.4.11",
    "uvicorn>=0.40.0",
]
